Consistently handle any messages
Fixed inconsistencies with regards to the existence checks and whether stderr, stdout or both should be discarded. In all cases, both should be discarded as the statements are to only check if the application exists. We are only interested in the return code and not any output that might be produced.

For the actual commands whether these should return messages (in particular error messages) is dependant on the designed purpose. If this is to only be executed by shortcuts and a terminal will never be seen then it is safe to assume that any form of message will not be needed. However, if this is intended to run in the console, I would remove all `>/dev/null 2>&1` or at least the stderr redirection so that in the case of failure a message explaining the failure is shown.

Finally on this point, I feel that the last commands in each case should be preceded with the `exec` command. The return code of `lxlock` should (in my personal opinion) not return 0 if in fact there was a failure to lock.

If either or both of these latter points seem reasonable, I will implement and submit another PR.