lxsession-logout: Avoid costly lsb_release invocation
Having the name of the distribution in the LXDE prompt looks kinda nice,
but `lsb_release -r -s` is fairly costly (or at least the version on
Debian is): It spins up a python interpreter, then runs `apt-cache policy`
(which in turn, among other things, runs
`/usr/bin/dpkg --print-foreign-architectures` twice).

On a desktop machine with a proper Haswell CPU, `lsb_release -r -s` takes
"only" around 105ms (around 79ms of that are for `apt-cache policy`);
but on an old Intel NUC with a 5-years-old mobile Intel Pentium with a TDP
of 6W, it takes roughly between 500ms and 1000ms.

As suggested by ZanderBrown, use glib's g_get_os_info() instead if
available. If that doesn't exist, it's probably better to omit the OS name
instead of potentially spending up to a second on figuring out what it is.