Remove `-d` option for `i3lock` in `lxlock` script.
Excerpt from i3lock man page:

"The  -d  (--dpms)  option  was  removed from
i3lock in version 2.8. There were plenty  of
use-cases  that were not properly addressed,
and plenty of bugs surrounding that feature.
While features are not normally removed from
i3 and its tools, we felt the need  to  make
an exception in this case."